% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_tests.R
\name{execute_tests}
\alias{execute_tests}
\title{execute tests}
\usage{
execute_tests(testdir)
}
\arguments{
\item{testdir}{directory under which all tests are stored}
}
\value{
named list of testresults
}
\description{
execute tests
}
