% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettests.R
\name{gettests}
\alias{gettests}
\title{extract tests from tarballs}
\usage{
gettests(tarbs, testDB = NULL)
}
\arguments{
\item{tarbs}{list of package source tarballs
from which tests have to be extracted}

\item{testDB}{directory under which package tests will be stored}
}
\value{
named vector of package test locations
}
\description{
extract tests from tarballs
}
